# static/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "VulnBank API",
    "description": "API documentation for the VulnBank banking application.",
    "version": "1.0.0",
    "contact": {
      "name": "VulnBank Support",
      "url": "https://github.com/Commando-X/vuln-bank"
    }
  },
  "servers": [
    {
      "url": "https://vulnbank.org",
      "description": "Controlled Production Server"
    }
  ],
...
  "paths": {
...
    "/login": {
      "post": {
        "tags": [
          "authentication"
        ],
        "summary": "Log in to the application",
        "description": "Authenticate user and receive JWT token.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string",
                    "example": "testuser"
                  },
                  "password": {
                    "type": "string",
                    "example": "password123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful login",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "token": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/transfer": {
      "post": {
        "tags": [
          "transactions"
        ],
        "summary": "Transfer money",
        "description": "Transfer money between accounts.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "to_account": {
                    "type": "string"
                  },
                  "amount": {
                    "type": "number"
                  },
                  "description": {
                    "type": "string",
                    "required": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Transfer successful"
          }
        }
      }
    },
...
    "/check_balance/{account_number}": {
      "get": {
        "tags": [
          "transactions"
        ],
        "summary": "Check account balance",
        "description": "Check balance of any account.",
        "parameters": [
          {
            "name": "account_number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Balance retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "username": {
                      "type": "string"
                    },
                    "balance": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Account not found"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT implementation"
      },
      "MerchantApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Merchant-Api-Key",
        "description": "Merchant API key returned by merchant registration/login"
      }
    }
  }
}